update the service to do async work in parallel
Currently the Google Maps API was initialized first and only then we asked
the user for their location. But this can be done in parallel. We don't
have to wait for Maps to be loaded to prompt the user to accept to share
their location.

Also, this commit improves the documentation of the service by quite a bit
and moves to code around into more sensible pieces.